wip: refactor to a module file
